requirements.txt
flask>=3.0
yt-dlp>=2024.1.1
pywebview>=5.0
